fix get_fname invalid read
