+++ interpretator/includes/interpretator.h
@@ -43,6 +43,11 @@ pid_t				g_job_pgid;
 int					g_shell_terminal;
 int					g_last_exit_status;
 pid_t				g_shell_pgid;
+
+/*
+**					find_in_path.c
+*/
+char				*find_in_path(char *filename, char **environ);
 /*
 **					exec_jobs.c
 */
@@ -113,7 +118,6 @@ void				open_pipe(t_process *p, int *pipes);
 /*
 **					get_fname.c
 */
-char				*find_in_path(char *filename, char **environ);
 int					check_path_var(char *fname, char **environ);
 char				*get_fname(char *arg);
 /*
