+++ interpretator/srcs/get_fname.c
@@ -12,49 +12,6 @@
 
 #include "interpretator.h"
 
-static char		*locate(char *fname, char **path)
-{
-	DIR				*dr;
-	struct dirent	*dirent;
-
-	if (!path)
-		return (fname);
-	while (*path)
-	{
-		if (!(dr = opendir(*path)))
-		{
-			path++;
-			continue;
-		}
-		while ((dirent = readdir(dr)))
-		{
-			if (!(ft_strcmp(dirent->d_name, fname)))
-			{
-				closedir(dr);
-				return (ft_pathjoin(*path, dirent->d_name));
-			}
-		}
-		closedir(dr);
-		path++;
-	}
-	return (NULL);
-}
-
-char			*find_in_path(char *filename, char **environ)
-{
-	char	**paths;
-	char	*res;
-	char	*path_var;
-
-	if (!(path_var = ft_getvar("PATH", environ)))
-		return (NULL);
-	if (!(paths = ft_strsplit(path_var, ':')))
-		err_exit("42sh", "malloc() error", NULL, NOERROR);
-	res = locate(filename, paths);
-	ft_free_dar(paths);
-	return (res);
-}
-
 static int		find_path_var(char *name, char **paths)
 {
 	char	*tmp;
